feat: user dropdown in friend list
