+++ src/frontend/app/src/types/StatusTypes.ts
@@ -1,0 +1,6 @@
+import type { UserState } from "./UserType";
+
+export interface StatusUpdate {
+	userId: number;
+	newState: UserState;
+}
