+++ src/frontend/app/src/types/UserType.ts
@@ -1,3 +1,14 @@
+export type UserState =
+  | "OFFLINE"
+  | "ONLINE"
+  | "CONNECTION_DENIED"
+  | "SEARCHING"
+  | "CHALLENGING"
+  | "PLAYING"
+  | "OBSERVING"
+  | "CHALLENGED";
+export type AuthenticatedState = "AUTHENTICATED" | "2FA" | "OAUTH";
+
 // When requesting information about other users
 export interface PublicUser {
   id: number;
