+++ src/frontend/app/src/utils/StatusTracker.ts
@@ -1,6 +1,5 @@
 import { useSocketStore } from "@/stores/SocketStore";
 
-const socketStore = useSocketStore();
 const SUBSCRIBE_EVENT = 'subscribeStatusUpdate'
 const UNSUBSCRIBE_EVENT = 'unsubscribeStatusUpdate'
 
@@ -9,11 +8,13 @@ function eventName(id: number) {
 }
 
 export function trackUserStatus(id: number, listener: (...args: any[]) => void) {
+	const socketStore = useSocketStore();
 	socketStore.status!.on(eventName(id), listener);
 	socketStore.status!.emit(SUBSCRIBE_EVENT, { id });
 }
 
 export function stopTrackingUserStatus(id: number, listener: (...args: any[]) => void) {
+	const socketStore = useSocketStore();
 	socketStore.status!.removeListener(eventName(id), listener);
 	socketStore.status!.emit(UNSUBSCRIBE_EVENT, { id });
 }
